I used Wireshark to record the HTTP traffic involved in creating a referendum with Helios. I then copy pasted the relevant flow (ignore css, js, png etc..) into helios-create-referendum.http.
The script create_referendum.py is a python script to programmatically
create a new referendum in Helios by doing HTTP GET and POST requests.
It can be used as a basis to make Helios a secure voting backend for
Solon.

Currently I have reverse engineered Helios "api" until the point
of freezing the election, which means users can go and vote.
Closing the election and computing the tally is still todo.